requirements.txt
crewai
streamlit
yfinance
python-dotenv
pandas
requests
duckduckgo-search
